requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==2.0.12
cloud-tpu-client==0.10
dj-database-url==2.1.0
Django==3.2.9
django-anymail==10.2
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
gcloud==0.18.3
google-api-core==1.34.1
google-api-python-client==1.8.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
h5py==3.10.0
httplib2==0.22.0
idna==3.6
Jinja2==3.1.3
jws==0.1.3
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.6.4
PyJWT==2.8.0
pyparsing==3.1.1
# Pyrebase==3.0.27
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2024.1
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
#tensorflow-macos==2.15.0
termcolor==2.4.0
torch==2.2.0
typing_extensions==4.9.0
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
